+++ src/graphnet/models/components/attention_blocks.py
@@ -1,5 +1,8 @@
 """Attention and transformer blocks used in graphnet models."""
 
+from typing import Optional
+
+import torch.nn as nn
 from torch.functional import Tensor
 from pytorch_lightning import LightningModule
 
@@ -34,3 +37,47 @@ class DropPath(LightningModule):
     def extra_repr(self) -> str:
         """Return extra representation of the module."""
         return "p={}".format(self.drop_prob)
+
+
+class Mlp(LightningModule):
+    """Multi-Layer Perceptron (MLP) module."""
+
+    def __init__(
+        self,
+        in_features: int,
+        hidden_features: Optional[int] = None,
+        out_features: Optional[int] = None,
+        activation: nn.Module = nn.GELU,
+        dropout_prob: float = 0.0,
+    ):
+        """Construct `Mlp`.
+
+        Args:
+            in_features: Number of input features.
+            hidden_features: Number of hidden features. Defaults to None.
+                If None, it is set to the value of `in_features`.
+            out_features: Number of output features. Defaults to None.
+                If None, it is set to the value of `in_features`.
+            activation: Activation layer. Defaults to `nn.GELU`.
+            dropout_prob: Dropout probability. Defaults to 0.0.
+        """
+        super().__init__()
+        if in_features <= 0:
+            raise ValueError(
+                f"in_features must be greater than 0, got in_features "
+                f"{in_features} instead"
+            )
+        out_features = out_features or in_features
+        hidden_features = hidden_features or in_features
+        self.input_projection = nn.Linear(in_features, hidden_features)
+        self.activation = activation()
+        self.output_projection = nn.Linear(hidden_features, out_features)
+        self.dropout = nn.Dropout(dropout_prob)
+
+    def forward(self, x: Tensor) -> Tensor:
+        """Forward pass."""
+        x = self.input_projection(x)
+        x = self.activation(x)
+        x = self.output_projection(x)
+        x = self.dropout(x)
+        return x
